requirements.txt
# API dependencies
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.18
sqlalchemy-utils==0.41.1
alembic==1.11.1
asyncpg==0.28.0
psycopg2-binary==2.9.6
pydantic==2.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.0.0.post2
prometheus-client==0.17.1

# Bot dependencies
aiogram==3.0.0
aiohttp==3.8.5

# Telegram API
telethon==1.29.2

# Task processing
redis==4.6.0
celery==5.3.1

# Security
cryptography==41.0.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3
aiofiles==23.1.0

# Development and testing
pytest==7.4.0
pytest-asyncio==0.21.1
black==23.7.0
isort==5.12.0
flake8==6.0.0